Chore(LOCAT-BE-37-pushNotification): refactor endpoint class
- endpoint 패키지 하위 클래스 user 패키지로 이동
- 예외 분리
