+++ src/main/java/com/locat/api/domain/user/controller/UserEndpointController.java
@@ -1,8 +1,8 @@
-package com.locat.api.domain.endpoint.controller;
+package com.locat.api.domain.user.controller;
 
 import com.locat.api.domain.core.BaseResponse;
-import com.locat.api.domain.endpoint.dto.EndpointRegistrationRequest;
-import com.locat.api.domain.endpoint.service.PlatformEndpointService;
+import com.locat.api.domain.user.dto.EndpointRegistrationRequest;
+import com.locat.api.domain.user.service.UserEndpointService;
 import com.locat.api.global.auth.LocatUserDetails;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -16,15 +16,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("v1/endpoints")
-public class EndpointController {
+public class UserEndpointController {
 
-    private final PlatformEndpointService platformEndpointService;
+    private final UserEndpointService userEndpointService;
 
     @PostMapping
     public ResponseEntity<BaseResponse<Void>> register(
             @AuthenticationPrincipal LocatUserDetails userDetails,
             @RequestBody @Valid final EndpointRegistrationRequest request) {
-        this.platformEndpointService.register(request, userDetails);
+        this.userEndpointService.register(request, userDetails);
 
         return ResponseEntity.ok(BaseResponse.ofEmpty());
     }
