+++ src/main/java/com/locat/api/domain/user/service/impl/UserEndpointServiceImpl.java
@@ -1,9 +1,13 @@
 package com.locat.api.domain.user.service.impl;
 
+import com.locat.api.domain.user.dto.EndpointRegistrationRequest;
 import com.locat.api.domain.user.entity.PlatformType;
 import com.locat.api.domain.user.entity.User;
 import com.locat.api.domain.user.entity.UserEndpoint;
+import com.locat.api.domain.user.service.PlatformEndpointService;
 import com.locat.api.domain.user.service.UserEndpointService;
+import com.locat.api.domain.user.service.UserService;
+import com.locat.api.global.auth.LocatUserDetails;
 import com.locat.api.infrastructure.repository.user.UserEndpointRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,6 +21,24 @@ import java.util.List;
 public class UserEndpointServiceImpl implements UserEndpointService {
 
     private final UserEndpointRepository userEndpointRepository;
+    private final PlatformEndpointService platformEndpointService;
+    private final UserService userService;
+
+    @Override
+    public void register(EndpointRegistrationRequest request, LocatUserDetails userDetails) {
+        List<UserEndpoint> endpoints = this.findUserEndpointsByUserId(userDetails.getId());
+        boolean endpointExists = this.isEndpointExists(request, endpoints);
+
+        if (!endpointExists) {
+            String endpointArn = this.platformEndpointService.create(
+                    request.deviceToken(), request.platform());
+            String subscriptionArn = this.platformEndpointService.subscribeToTopic(endpointArn);
+
+            User user = this.userService.findById(userDetails.getId());
+            this.saveUserEndpoint(
+                    user, request.deviceToken(), request.platform(), endpointArn, subscriptionArn);
+        }
+    }
 
     @Override
     public List<UserEndpoint> findUserEndpointsByUserId(Long userId) {
@@ -37,5 +59,11 @@ public class UserEndpointServiceImpl implements UserEndpointService {
                 .build();
 
         this.userEndpointRepository.save(userEndpoint);
+    }
+
+    private boolean isEndpointExists(EndpointRegistrationRequest request, List<UserEndpoint> endpoints) {
+        return endpoints.stream()
+                .anyMatch(e -> e.getDeviceToken().equals(request.deviceToken())
+                        && e.getPlatformType().getValue().equals(request.platform()));
     }
 }
